refactor: move sample CSV utility to lib
