+++ src/features/deck-import/ui/DeckImportView.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 import { INITIAL_VIEWPORTS } from "@/storybook/storybookViewports";
 
 import type { DeckImportPreview } from "../hooks/useDeckImport";
-import { SAMPLE_CSV_TEXT } from "../sampleCsv";
+import { SAMPLE_CSV_TEXT } from "../lib/sampleCsv";
 import { DeckImportView as Template } from "./DeckImportView";
 
 const preview = {
